+++ database/gdb/gdb_core_underlying.go
@@ -396,7 +396,9 @@ func (c *Core) RowsToResult(ctx context.Context, rows *sql.Rows) (Result, error)
 		record := Record{}
 		for i, value := range values {
 			if value == nil {
-				record[columnNames[i]] = gvar.New(nil)
+				// Do not use `gvar.New(nil)` here as it creates an initialized object
+				// which will cause struct converting issue.
+				record[columnNames[i]] = nil
 			} else {
 				var convertedValue interface{}
 				if convertedValue, err = c.db.ConvertValueForLocal(ctx, columnTypes[i], value); err != nil {
